Update: Based on the comments
