+++ backend/app/Http/Controllers/Auth/RegisterController.php
@@ -10,7 +10,7 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Validation\Rules;
 
-class RegisteredUserController extends Controller
+class RegisterController extends Controller
 {
     /**
      * Handle an incoming registration request.
