+++ backend/app/Http/Middleware/VerifyEmail.php
@@ -5,7 +5,7 @@ namespace App\Http\Middleware;
 use Closure;
 use Illuminate\Contracts\Auth\MustVerifyEmail;
 
-class EnsureEmailIsVerified
+class VerifyEmail
 {
     /**
      * Handle an incoming request.
